% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS.R
\name{LS}
\alias{LS}
\title{Fixed effect meta-analysis for correlated test statistics}
\usage{
LS(beta, stders, cor = diag(1, length(beta)))
}
\arguments{
\item{beta}{regression coefficients from each analysis}

\item{stders}{standard errors corresponding to betas}

\item{cor}{correlation matrix between of test statistics.  Default considers uncorrelated test statistics}
}
\value{
Return values:
\itemize{
\item{beta: }{effect size}
\item{se: }{effect size standard error}
\item{p: }{p-value}
}
}
\description{
Fixed effect meta-analysis for correlated test statistics using the Lin-Sullivan method.
}
\details{
Perform fixed effect meta-analysis for correlated test statistics using method of Lin and Sullivan (2009).  By default, correlation is set to identity matrix to for independent test statistics.
}
\examples{
# Generate effects
library(mvtnorm)
library(clusterGeneration )

n = 4
Sigma = cov2cor(genPositiveDefMat(n)$Sigma)
beta = t(rmvnorm(1, rep(0, n), Sigma))
stders = rep(.1, n)

# Run fixed effects meta-analysis, 
# assume identity correlation  
LS( beta, stders)

# Run random effects meta-analysis,
# assume identity correlation  
RE2C( beta, stders)

# Run fixed effects meta-analysis, 
# account for correlation 
LS( beta, stders, Sigma)

# Run random effects meta-analysis,
# account for correlation 
RE2C( beta, stders, Sigma)

}
\references{
{
  \insertRef{lin2009meta}{remaCor}
}
}
