% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metanova.R
\name{metanova}
\alias{metanova}
\title{Meta-analysis on multivariate regression}
\usage{
metanova(fit, method = c("LS", "RE2C"))
}
\arguments{
\item{fit}{multivariate regression object from lm()}

\item{method}{specify the test: "LS" or "RE2C"}
}
\description{
Meta-analysis on multivariate regression using either LS (fixed) or RE2C (random test)
}
\examples{
# simulate data
N = 100
Y = cbind(rnorm(N), rnorm(N))
x = rnorm(N)

# fit multivariate regression
fit = lm(Y ~ x)

# evaluate LS test
metanova(fit, "LS")

# evaluate RE2C test
metanova(fit, "RE2C")

# standard anova test
anova(fit)

}
