% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metanova.R
\name{print,metanova-method}
\alias{print,metanova-method}
\title{print for metanova}
\usage{
\S4method{print}{metanova}(x, ...)
}
\arguments{
\item{x}{metanova object}

\item{...}{other arguments, currently ignored}
}
\value{
results of print
}
\description{
print for metanova
}
