% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metanova.R
\name{show,metanova-method}
\alias{show,metanova-method}
\title{show for metanova}
\usage{
\S4method{show}{metanova}(object)
}
\arguments{
\item{object}{metanova object}

\item{...}{other arguments, currently ignored}
}
\value{
results of show
}
\description{
show for metanova
}
